Revert to using two functions for requesting VCF and TBI files.
VCF and TBI files need to be specified in merge so they are
produced in the scratch subdirectory. Generation of these files
depends on the main_chromosomes.txt to be generated. If
main_chromosomes.txt has not been generated, a function which
returns VCF and TBI files as a named dictionary will not return
the dictionary and leaving the input of _merge_vcfs to be
without key:name values. Any references to these keys in
_merge_vcfs shell directive will cause a key name error.